Refactor classpath URL handler for easier Java 8 support
